soc: arm: infineon_xmc: Add support for using cachable flash address
The infineon xmc4xxx series has two ways to access flash: One is the
cachable address space at 0x8000000 which can be used to improve read
access; The second is uncached space at 0xc000000 which is mainly used
for write and erase operations.

This patch adds support for using the cachable read access by setting
it as the virtual address in the linker script for the rom section.

Signed-off-by: Andriy Gelman <andriy.gelman@gmail.com>